4.21 exercises
Change the delete blog operation so that a blog can be deleted only by the user who added the blog

+++ part4/controllers/blogs.js
@@ -63,8 +63,18 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-	  await Blog.findByIdAndRemove(request.params.id)
-	  response.status(204).end()
+	const decodedToken = jwt.verify(request.token, process.env.SECRET)
+	if (!decodedToken.id) {    
+		return response.status(401).json({ error: 'token missing or invalid' })  
+	}
+	const blog = await Blog.findById(request.params.id)
+	if(blog.user.toString() === decodedToken.id){
+		await Blog.findByIdAndRemove(request.params.id)
+	    response.status(204).end()
+	}else{
+		return response.status(401).json({ error: 'Unauthorized to delete the blog' })
+	}
+	  
   })
